Refactored tests, added 2 tests on invocation parameters of page extension
