+++ test/cypress/integration/dialog-extension.spec.js
@@ -14,7 +14,7 @@ import { verifyLocation } from '../utils/verify-location'
 import {
   verifySdkInstallationParameters,
   verifySdkInvocationParameters
-} from './reusable/open-sdk-parameters-test'
+} from '../utils/verify-parameters'
 
 const post = {
   id: '3MEimIRakHkmgmqvp1oIsM',
@@ -62,7 +62,7 @@ context('Dialog extension', () => {
     cy.get('@extension').should('be.visible')
   })
 
-  it('sdk.ids static methods have expected values', () => {
+  it('verifies sdk.ids static methods have expected values', () => {
     cy.getSdk(iframeDialogSelector).then(sdk => {
       expect(sdk.ids.contentType).to.equal(undefined)
       expect(sdk.ids.entry).to.equal(undefined)
@@ -80,7 +80,7 @@ context('Dialog extension', () => {
     })
   })
 
-  it('sdk.parameters have expected values', () => {
+  it('verifies sdk.parameters have expected values', () => {
     cy.getSdk(iframeDialogSelector).then(sdk => {
       verifySdkInstallationParameters(iframeDialogSelector)
       verifySdkInvocationParameters(iframeDialogSelector)
